Unnecessary classes removed
The original logic was this:
- I have N dispatchers that listens to their assigned queues/tubes and;
- as soon as they receive/reserve a message, if there are enough free
slots (this check is done synchronously by the
RunningProcessHelper):
- create a new thread and run the received command;
- when done, delete the message on the queue.

Problem: JobIDs are local to the current client connection (ie: I cannot
use a new client in order to delete messages that have been reserved by
another one).
So, a simpler approach has been used: the Clidispatcher waits for the
answer by the runned command, and when it's done, it deletes the
message.
The Apache Commons Exec library Watchdog ensures that the process does
not go beyond the maximum timeout. If everything explodes, the
beanstalkd
server itself, releases messages after the timeout.

More info in the code.